Extract LayerConstellationsModule
