+++ src/app/app.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { AstrocadreComponent } from 'src/app/components/astrocadre/astrocadre.component';
+import { LayerConstellationsModule } from '#layer-constellations/layer-constellations.module';
 import { CoreModule } from '#core/core.module';
 import { ControlsModule } from '#controls/controls.module';
 import { LayerSkyGridModule } from '#layer-sky-grid/layer-sky-grid.module';
@@ -17,7 +18,8 @@ import { LayerSkyGridModule } from '#layer-sky-grid/layer-sky-grid.module';
   imports: [
     CoreModule,
     ControlsModule,
-    LayerSkyGridModule // TODO remove when loaded dynamically
+    LayerSkyGridModule, // TODO remove when loaded dynamically
+    LayerConstellationsModule
   ],
   bootstrap: [AstrocadreComponent]
 })
