+++ src/app/modules/layer-constellations/layer-constellations.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CoreModule } from '#core/core.module';
+
+@NgModule({
+  imports: [
+    CoreModule
+  ]
+})
+export class LayerConstellationsModule {
+
+}
