+++ src/app/modules/layer-constellations/models/constellation-lines-layer-factory.ts
@@ -1,5 +1,5 @@
 import { Layer } from '#core/models/layers/layer';
-import { ConstellationLines } from '#core/models/layers/constellation-lines';
+import { ConstellationLines } from '#layer-constellations/models/constellation-lines';
 import { AxialCurvesFactory } from '#core/models/layers/factories/axial-curves-factory';
 import { LayerFactory } from '#core/models/layers/factories/layer-factory';
 import { SupportedLayers } from '#core/models/layers/supported-layers';
