+++ src/app/modules/layer-constellations/models/constellation-lines.ts
