+++ tsconfig.json
@@ -30,6 +30,9 @@
       "#core/*": [
         "src/app/modules/core/*"
       ],
+      "#layer-constellations/*": [
+        "src/app/modules/layer-constellations/*"
+      ],
       "#layer-sky-grid/*": [
         "src/app/modules/layer-sky-grid/*"
       ],
